wayland: Add xx-session-management-v1 support
This is powered by the xx-session-manager which is tagged as
experiemntal. This is guarded by an env variable.

Application level session IDs are exposed via QSessionManager.

The public API for windows is exposed via the existing QSessionManager
and QWidget::setWindowRole.

As it's not widely supported the documentation in QWidget is left
untouched for now.

[ChangeLog][Third-Party Code] New protocol synced from wayland-protocols

Change-Id: Ibfbef86c6e75f8b95433cbba69ca10a5abea9e21
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>