Doc: Fix issues in global documentation template
The qt-html-templates-offline.qdocconf includes
html-header-offline.qdocconf, which defines `HTML.stylesheets`,
the files to copy to /style subfolder in the output directory.

This variable was then later incorrectly redefined directly in
qt-html-templates-offline, missing one file, `tech_preview.svg`.
This file is used in CSS only, not in documentation sources, so
QDoc did not warn about the omission.

Remove duplicate definitions, and move all variables that do not
relate to HTML headers away from html-header-offline.

In passing, clean up formatting of the related definitions and
fix incorrect paths passed to `DocBook.extraimages'.

Pick-to: 6.10 6.8
Change-Id: I2428aaa8eb5871ec9757f74695096bc42664f1f2
Reviewed-by: Topi Reinio <topi.reinio@qt.io>