wasm: Introduce wasmdeployqt
Introduce wasmdeployqt which is analogous to other
deployment tools that we currently have like
"windeployqt" or "androiddeployqt".
Running wasmdeployqt makes it much easier to deploy
and ship dynamically linked WebAssembly applications.
In case of statically linked applications this tool
does not provide any extra value as static apps are
ready to be shipped by default.

This commit also removes older python scripts which
were supposed to serve this same goal as well as
some cmake machinery arround it.

Fixes: QTBUG-134580
Change-Id: Ib5c8ab128f49897f91c510bfec03b0b772e6548a
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>