QVarLengthArray: improve assign() for forward iterators (typical case)
This disentangles the forward iterator case from the more complex,
iterative input-only iterator one, because a) we can know the final size
from the beginning and b) we can use the the std copy_n algorithms. This
generates far better code for the more common uses of random-access
iterators and especially that of trivial types.

Pick-to: 6.10
Change-Id: Idee20c7c877884baf9befffd5e00b576ebfa2f86
Reviewed-by: Marc Mutz <marc.mutz@qt.io>