QMenu: Fix transient parent for nomadic menus
There are applications that reuse menus for different user interface
elements. For example, a menu can be shown in the menubar or in a right
click context menu.

On the other hand, the QMenu currently locks on the first transient
parent and doesn't change it even after effectively getting attached to
another parent.

This change addresses the issue by making the QMenu re-evaluate the
transient parent on every show event and also ensuring that the caused
widget takes the precedence over other heuristics. The caused widget is
more closely related to the user input.

Pick-to: 6.9 6.10
Fixes: QTBUG-137467
Change-Id: I4f968833380121be1300af1a44237735b9a97b17
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>