Android: auto-tests: Add missing lib to tst_qapplication Android apk
tst_qapplication is missing the modal_helper executable from
the apk. Set the libmodal_helper.so as a target property for the
tst_qapplication with QT_ANDROID_EXTRA_LIBS property.

Change the name of qtbug_12673() test function to modalDialog(),
I think this better represents what's being tested.
The bug ticket representing qtbug_12673() can be found from
the ticket linked to this commit.

Construct a full path to the modal_helper.so and pass that
to QProcess.start instead of relative path to filename
when targeting Android.

Add a shared utility function androidAbi() that returns
the currently defined Android ABI.

Change the function name in BLACKLIST file.

Task-number: QTQAINFRA-6908
Pick-to: 6.8 6.9 6.10
Change-Id: I13904acda0f5608ea31df49bd95824e1412f2786
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>