QFontIconEngine: use the horizontalAdvance when calculating actualSize
Glyphs used for the icon might have left or right bearing, and we need
to include those metrics in the calculation of the actual size of the
icon so that QIcon correctly aligns it in the target rectangle when
painting.

Observed on Windows with the icon used for
QIcon::ThemeIcon::DialogInformation; the glyph for \ue946 in the
Segoe Fluent Icons font has a right bearing of 18.75, and the icon was
rendered too far to the left, truncating pixels.

Using the horizontalAdvance instead of the boundingRect includes those
values. QFontMetrics doesn't have an API that give us the minimal rect
that fits all pixels, or the actual height of a specific text, so use
tightBoundingRect.height for the height.

Fixes: QTBUG-136176
Fixes: QTBUG-134604
Pick-to: 6.10 6.9
Change-Id: I7c0065c625aa9b01f8eaf770ba13a0995f3df4de
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>