wayland: Move error checking before xdg surface creation
8c0dd12f4bf7a09304a136c9ceacfb60e0632e50 changed behaviour when
QtWayland is asked to enter an invalid state of creating a popup where
no parent window is explicitly set or can be guessed. It now early
returns before creating an xdg role.

It is apparently an error on some compositors to create and then destroy
an xdg_surface without creating a role.

Fixes: QTBUG-137702
Pick-to: 6.10
Change-Id: I8b4dcb54f03660cb47e93b54a68834b371f76c71
Reviewed-by: David Redondo <qt@david-redondo.de>