QDBusMessage: add move-constructor
Since the class was not desigend to support a nullptr d_ptr, this
change requires modifications in the destructor and copy-ctor.
The change in destructor is straightforward - simply add a check
that d_ptr is not null.
The copy-constructor was using a qAtomicAssign() helper, which
was relying on the fact that the passed pointers are not null, so
we cannot use it anymore. Use copy-and-swap instead.

The other methods do not require any changes, because the
moved-from object can only be destroyed or assigned-to.

[ChangeLog][QtDBus][QDBusMessage] Added move constructor.

Change-Id: Ic8a0d913b9cf2f881369f7ad4f3a88c1f3fb345f
Reviewed-by: Marc Mutz <marc.mutz@qt.io>