Mark qwindowspipereader.cpp as security-critical
QWindowsPipeReader is a candidate for a security-critical class because
it is parsing data from an external process. The parsing is trivial
(e.g. searching \n in *.cpp line 243 or skipping characters) and
the data should arguably come from trusted sources. We still go for the
safe option and mark it as critical.

QUIP: 23
Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-135187
Change-Id: I79767534015370a5102929fee032b916be23d41d
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>