windeployqt: Fix access of invalid iterators crash
We must not use a range based loop here as the container might be
changed inside the loop which will invalidate the iterators which in
turn might lead to a crash.

While at it the QFileInfo used to obtain the binary's path was moved out
of the loop as a small cleanup.

Fixes: QTBUG-137763
Pick-to: 6.10 6.9 6.8
Change-Id: I86fb8b683fb98a897d0059ef1dc869e26b2c213f
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>