Specify CMAKE_MAKE_PROGRAM explicitly
This works around the issue that we have when using vcpkg. It's
internal logic adds system root to CMAKE_PREFIX_PATH when
crosscompiling. As result ninja executable that is located by PATH
is ignored and the system one is used. There is no way to disable
this behavior from vcpkg side, so we need to set the CMAKE_MAKE_PROGRAM
variable explicitly.

Change-Id: I18fb41a525d6ddc70b0e8df12426146f2051e904
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>