Android: clean up Gradle build after APK creation to save space
Gradle builds take quite a lot of space for each project, and that is
even higher when using a -developer-build. So, running multiple/all
tests on a local machine most of the time ends up filling the whole
disk. To avoid this, this adds QT_ANDROID_POST_BUILD_CLEANUP as
a CMake option to opt-in for running 'gradlew clean' right after the
APK has been created, keeping only the Gradle project structure and
the APK which don't take as much space.

Task-number: QTBUG-94777
Change-Id: I7184157909934e65d8d73ecca168e26979dea3f3
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>