iOS: implement file icon provider
Move the existing QCocoaFileIconEngine into a separate file in the
darwin platform code, from where we can use it from both the cocoa
and the iOS theme.

Refactor the implementation for macOS to create and retain the NSImage
as a member of the engine, and cache the QPixmap when it's requested,
reusing it as long as it has the correct size.

The iOS implementation is similar, except we need to go through
UIDocumentInteractionController to get the icons for the file's URL.
Explicitly make sure that we maintain the aspect ratio of the image we
get.

Augment the iconbrowser manual test to generate temporary files with
certain extensions, and include those in the UI, allowing us to test
that the icons we get from QAbstractFileIconProvider match the file
type.

Fixes: QTBUG-134239
Change-Id: I8fb63b3c518a6eb200f5948a1c38fd485e3b1c6d
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>