QMetaType::fromName: move normalization code up to fromName() itself
That way, we can have a single implementation that is no longer a
template, shared between fromName() and qMetaTypeTypeInternal().

This does mean we will unlock and relock the registry in case the type
normalization produced something different. But it does mean we won't
run the normalization with the registry locked.

Pick-to: 6.10
Change-Id: Ic704942a45cea90a17fefffdcbb0cb9573bdab03
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>