Apple icon engine: define more symbols
On iOS, the native icon engine is used for the standard icons of
QAbstractFileIconProvider. For File, we map to text-x-generic, and for
Network to network-workgroup. Map those to sensible icons from SF
Symbols.

For the Desktop icon, switch to the same icon that Finder is using for
that.

Task-number: QTBUG-134239
Pick-to: 6.10
Change-Id: I962809543bcbd1d900c1083ff0bb85b9b7f162f8
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>