PSL update script: Change to use direct upstream source
The upstream "Public Suffix List" .dat file has gained a comment line
with a "// VERSION:" field, so the work to extract a git SHA is not
needed, nor is it canonical with their new versioning. Though they also
embed the SHA from which the file originated, which we continue to use
for the PURL.

Furthermore the file now warns that any other source of the PSL other
than the one from publicsuffix.org is "not guaranteed to be supported".
So in this patch we move to use their canonical upstream http download:

https://publicsuffix.org/list/public_suffix_list.dat

Pick-to: 6.10 6.9 6.8 6.5 5.15
Task-number: QTBUG-137782
Change-Id: I50ec8ed6a55738560d9108ad08517c243c1bb518
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>