QMetaType: de-virtualize the QMetaTypeModuleHelper struct
Now instead of storing a pointer to a vtable, we store the vtable
directly. That is one fewer pointer dereference to get to where we want
to go. It does mean you can't press F2 in Qt Creator to find the other
implementations (not that "._anon_4" was of any help).

Change-Id: I5635ebd8907c79c2a935fffdfc8952246ed7fb04
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>