QXcbWindow: Lock m_mapped, when applying state to unmapped window
If QXcbWindow::setNetWmStateOnUnmappedWindow() is called on a window
that has already been requested to be mapped, it can be mapped before
the state has been applied, leading to the window having the wrong
state (e.g. requested modality not respected).

Eliminate the data race by locking m_mapped with the existing mutex
member.

Fixes: QTBUG-120031
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I2b142420fc155d8b465df846560d868bd87882fb
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>