Mark QFile security significant
QFile is realativly simple because most of the work is forward to other
classes and files.

The opening of existing file handles looks a bit sketchy and we don't
test for e.g. FILE *fh being a nullptr. But that should really be
covered by the application developer.

QUIP: 23
Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-135187
Change-Id: I926bc23859d47e6e8a9da542e0c3d5d39ab8f85b
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>