QXcbWindow: Eliminate data race between hide() and deferred activation
When hide() was called on a window that is still hidden with its
activation being deferred, the hide() call could have been overridden
when the deferred activation kicked in.
In tst_QGraphicsScene::tabFocus_sceneWithFocusWidgets() this sometimes
causes the wrong widget to be focused.

Lock the mutex for m_mapped at the beginning of hide().
Cancel deferred activations when setting m_mapped to false.

Can't be autotested.
Flakiness of tst_QGraphicsScene::tabFocus_sceneWithFocusWidgets() will
disappear.

Fixes: QTBUG-123063
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I80c3419f09a5fd98c2684c672c631c6a63d4ec48
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>