QRangeModel: documentation improvements
Some restructuring, and include a section on rows as values vs rows as
pointers. That paragraph as written so far was confusing, even for me,
and the subject matter is complex enough to require more than a single
paragraph anyway.

Still work in progress, esp with some pending changes to enforce the
requirement that ranges of pointers cannot be passed by value.

Pick-to: 6.10
Change-Id: Ie4496ffb32b2622835c3cc4662da129e34e3f8c7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>