wasm: move input event handlers to QWasmWindow
Attach input event handlers to the m_window elemement,
and make the input context input element be a child
element of m_window. Forward input and composition
events to QWasmInputContext when it is active.

Add a new focusHelper hidden child element which takes
focus whenever the window has focus and the input element
does not. Using a focus helper means we don't have to
set contentEditable on m_window and avoid the side effects
caused by that. Make the focus helper take focus on
window focus, instead of the canvas.

This setup ensures that the window can receive and
intercept input events during the event capturing phase,
regardless of which child event has focus.

This also allows for placing focus on the accessibility
elements; QWasmWindow can then decide if it wants to
intercept events or pass them through, based on event
type and target.

Pick-to: 6.10
Change-Id: I9ec6f8e8f7ec0878009eddd05642dc146ea3944f
Reviewed-by: Lorn Potter <lorn.potter@qt.io>