QArrayDataOps::assign: simplify undoing the prepend optim. for trivials
Since they are trivial, we don't need to loop destroying anything (dead
code anyway), so help the compiler out and always store the new capacity
begin. This may be storing the value that was already there if no
element was removed from the beginning of the list, but it does remove a
conditional.

Further, this removes the entire looping over the prepend optimization
buffer for new elements. We'll fall straight to the main assignment
loop.

Change-Id: I5338b6a53bffbda17a08fffd1d7eda72d57eff29
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>