QArrayDataOps::assign: improve for forward iterators (typical case)
Similar to commit 7ed2db674328f1c96d06be5ffb15b8ffefeae545, which did
this for QVarLengthArray. This disentangles the forward iterator case
from the more complex, iterative input-only iterator one, because we
know the final size from the beginning. This generates far better code
for the more common uses of random-access iterators and especially that
of trivial types.

Change-Id: Ib71e489ae0dafb7a917bfffd3441d8c14b82ad46
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>