cmake: remove variables that doesn't exist
This "deprecation_define" variable existed before, but after
some commit the logic that set this variable has all been
removed, so it's always empty and become useless now.
I've searched for this variable globally in whole qtbase repo
and can't find anywhere set or use this variable, so I think
it's safe to remove these leftovers.

Change-Id: If89c40570f24f5b845d17966da8a83746473152f
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>