Windows QPA: fix PMv2 detect logic
Before this patch, in QWindowsContext::setProcessDpiAwareness,
there are some conditions that will make this function return
early, in this case, QWindowsContextPrivate::m_v2DpiAware won't
be updated to the correct value (it's initially set to false
and we want it to be true in these cases), this is absolutely
wrong. Luckily, this issue just makes QPA call enableNonClientDpiScaling() in unnecessary situations and
that is harmless. But it's also better to fix this issue.

Change-Id: I04d70d5d83b898a3cd2856d504bcd3f5519427b4
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>