Mark QStandardPath implementation security critical
The QStandardPaths class provides access to OS-specific standard
locations on the filesystem, and also allows to search for files
within the defined directories. Other parts of the application
should be able to rely on the results returned by this class. Bugs
in this code can lead to severe issues like overwriting of user files,
reading configs from an incorrect place, or loading an incorrect
external code for execution. Therefore, the respective files are
marked security-critical.

QUIP: 23
Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-135187
Change-Id: Id1a8189f2a7cde4a760446428551f2fe682d4073
Reviewed-by: Marc Mutz <marc.mutz@qt.io>