Mark QStorageInfo security significant
QStorageInfo parses paths passed to QStorageInfo::setPath. We do
not call this function ourselves with untrusted data. It has to
be called by the application developer, and as such, it should
be their responsibility to validate the passed path.

Furthermore, this class parses various system files, which are
treated as trusted data sources.

QUIP: 23
Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-135187
Change-Id: I7e0cca60b69c6f9698442b19135c947fa95e62bf
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>