tst_QNetworkInterface: fix localAddress() for point-to-point connections
When the IPv4 netmask is 255.255.255.255 (point-to-point interface), we
were calculating our own local IP. That meant we would route through the
loopback interface to connect to it, not the interface, resulting in
test failures. So just flip one bit and hope that it will work.

FAIL!  : tst_QNetworkInterface::localAddress(vpn0-10.124.222.242) The computed value is expected to be less than or equal to the baseline, but is not
   Computed (pmtu)                                    : 65535
   Baseline (outgoingIface->maximumTransmissionUnit()): 1422

$ ip r get 10.124.222.242
local 10.124.222.242 dev lo table local src 10.124.222.242 uid 1000

Pick-to: 6.10 6.9 6.8
Change-Id: I3d18ad117dd0c49fcec2fffdcb29534e64a438d9
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>