QPainter: make qt_show_painter_debug_output constexpr
Nothing is modifying the variable, so make it constexpr (was: mutable)
and change all the if's into if-constexpr's.

While the compiler doesn't care unless QT_DEBUG_DRAW is defined, this
will prevent the next human reader from going on the same fruitless
hunt for a setter as yours truly.

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ib2f8c0fcb2938ca3548b9452c4d8eb482e6ffbe5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>