QLockFilePrivate: de-inline all functions
This class is going to be marked as security-critical, and so we don't
want code in the header (because fixes in inline code require
recompiling all users, not just re-linking).

Incidentally, de-inlining getLockFileHandle() removes the need to
include some nasty (qt_windows.h) headers in the header (which means
we need to revert part of 779bdf481ca208c20cfabd02d02d74175b0624b9 and
go back from NSDMI to ctor-init-list, at least for `fileHandle`).

As a drive-by, make the constructor explicit.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I3d376c2c02eb1f7de7a0b923ac4424b1057fdb1c
Reviewed-by: David Faure <david.faure@kdab.com>