QLockFile: make d_ptr private
The original author of the class doesn't remember why it's protected,
and there appear to be no meaningful uses for a protected d_ptr
(unlike in the QObject case). The class is also not designed as a
base class (the destructor is public and non-virtual), and we never
advertized the protectedness of the d_ptr, so treat this as \internal
API and just make it private.

[ChangeLog][Potentially Source-Incompatible Changes][QtCore] The
protected QLockFile::d_ptr variable is now private. The class isn't
safe for subclassing, anyway (doesn't have a virtual destructor), so
if your code depends on this, you need to find a different design.

Change-Id: Idba36fd65f4c1327f31ee6c27320a0697147de5f
Reviewed-by: David Faure <david.faure@kdab.com>