Mark QTextStream as security-critical
This component processes input external to the process, and the
application depends on this component to report errors to detect
failure, so this component is security-critical.

The headers, however, contain only a miniscule amount of
implementation, essentially a few forwarders each and, in the public
header, the QTextStreamManipulator, everything else is out-of-line, so
I've opted to mark both headers as significant, header,
declarations-only. I'll de-inline what makes sense in a follow-up
commit.

Amends 8df072fc8006510c9b743e8ffedaaf51a876883a.

This completes the review of src/corelib/serialization, building on
the following prequel commits:
- 0b9d4c3a857144bad2937f964101c57ae7efec4c (QDataStream)
- 0a7ee06b27c55f10b65b053cabdc893ae8f23893 (QXmlStream{Reader,Writer})
- 046e6c5701da96dd64b613c2bc9cfd245c83b1a5 (QXmlUtils)
- cd4813a3c304a6d9d5d8096d91a8fc297818d624 (QJson and QCbor)

There are no other implementation files in this subdir at the time of
commit.

QUIP: 23
Fixes: QTBUG-135194
Pick-to: 6.10 6.9 6.8
Change-Id: If5264e3ad1eca230f48042888b1256ce0a329d8e
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>