QObject: make doSetProperty() take lvalue arg by cref
... not by pointer.

It makes more sense this way, since `lvalue` is always passed (unlike
`rvalue`, which is passed optionally, so needs a nil state), and
removes a potential error state (lvalue == nullptr) that neither the
compiler nor a reader of the code now have to worry about anymore.

As a drive-by, rename the argument so the local alias in the function
can go away, too.

Amends 39cdf431f034121353e51768b4d1fec8b0dd35dc.

Pick-to: 6.10
Change-Id: I818e4fec30360c5b63b84a0f7e1e1eebdfe2cfcc
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>