QQuaternion: comment all #endif's
It's easier to comprehend what's going on, because there are quite a
few different ones here.

Exception: if the #if block contains only very few lines (max four or
so, didn't count), then a naked #endif is ok, and arguably easier on
the eye, so leave them in that case.

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ib1b27859344916caa17dc1ff0010b6b10d34bb13
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>