qmatrix4x4.h: don't include qquaternion.h
It's only used as an argument of an out-of-line function, so a forward
declaration suffices.

Found while wondering why a change to qquaternion.h recompiles most of
Qt ;-)

[ChangeLog][Potentially Source-Incompatible Changes][QtGui] The
qmatrix4x4.h header no longer includes qquaternion.h. A
backwards-compatible fix is to not rely on transitive includes and
include what you need explicitly.

Pick-to: 6.10
Change-Id: Ic0e28155efa0342f22b08da28776819694fddcb5
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>