QIdentityProxyModel: call data() from subclasses in match()
The implementation of match() was only looking into the source model,
without considering the fact that data() could be reimplemented in
a QIdentityProxyModel subclass (a very common usage pattern for the
class, part of the documentation).

The simplest solution is to let the default implementation from
QAbstractItemModel do matching the usual way, calling data().

Fixes: QTBUG-78013
Pick-to: 6.10 6.9 6.8
Change-Id: If984ffa5f8dbdfe9a76eaec8b82648d705cfecd4
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>