tst_QTemporaryFile::stressTest(): reliably remove temp files [1/2]: guard
Use a QScopeGuard to remove all the temp files created, so as to also
remove them if a QVERIFY should fail.

There's still a hole in the test: When the contains() QVERIFY fails,
we won't delete that temp file, but that will be solved by porting to
QDuplicateTracker in a subsequent commit.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ib3c4082c4ad11b271b29491e1be0b62b291684fa
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>