tst_QTemporaryFile::stressTest(): reliably remove temp files [2/2]: dup tracker
After the first patch already reliably removes the temp files left by
the test in most cases, this patch plugs the remaining hole, namely
that the insert() hasn't happened, yet, when the contains() QVERIFY()
fails, and we therefore don't QFile::remove() the youngest temp
file.

Use QDuplicateTracker, whose raison d'etre is that it has API that
does lookup and insert in one go, so that a failed QVERIFY will have
already added the file name to the container. We recently added
iteration support to QDuplicateTracker, making it possible to use it
here as a drop-in replacement for QSet.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ia6d1cfa50e452cf583935d5f76aa1cb6794d6eeb
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>