QDataBuffer: fix incorrect use of realloc()
One must never overwrite the first argument of realloc() with its
result, before checking the result for nullptr, otherwise the old
pointer value (which still needs to be free()d when realloc() fails)
is lost and therefore memory is leaked.

As a drive-by, replace C-style with static_cast.

Found by explicit review of realloc() calls (incl. QtPrivate
wrappers).

Amends the start of the public history.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ifac55291898837a9e063c508e03f1c7da0a0e61f
Reviewed-by: Aurélien Brooke <aurelien@bahiasoft.fr>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>