Correct Africa/El_Aaiun to fix typo in CLDR data
Amends commit 7176ca82bf5e69f58d863e738ca249fd91f64190 - there's a
stray &lt; in the upstream that still hasn't been fixed, so I've
corrected it locally and regenerated the data. I used the same fixed
data when regenerating 6.8's data to fix conflicts in the pick back to
it, so this doesn't need to be picked to 6.8.

This removes a spurious entry from exemplarCityTable[], at index
221031 of length 9, thereby shifting all later entries earlier by 9
characters.

Pick-to: 6.10 6.9
Task-number: QTBUG-137782
Change-Id: Iedbbd38e6703cf9152e17bd277a9e1b9c6fe3af6
Reviewed-by: Mate Barany <mate.barany@qt.io>