QtTest WatchDog: fix UB (default-constructed std::atomic)
Up to C++17 inclusive, a default-constructed std::atomic object can
only be initialized using std::atomic_init(), _not_ by .store().

While it's probably not a problem in practice, it's also not much work
to call the std::atomic<T>(T) ctor in the ctor-init-list instead of
the implicit default constructor, followed by a store() in the body of
the ctor, so do that.

Amends a8a38f7caabb90113a229629c6fe463acdecd10f.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I9ba953d1bc34fe09326554a65393b8a2deb35328
Reviewed-by: David Faure <david.faure@kdab.com>