Revert "wasm: remove unsupported libraries from wasm build"
This reverts commit 44b513f192ccd0fc25e8ba7bddad5d8323210a18.

Enable SQL and PrintSupport again; we support sqlite and print to pdf.

In addition we don't want to wasm to diverge from the standard
module set if possible, in order to avoid build system issues.
For example, disabling Qt::PrintSupport removes QT_FEATURE_printer,
which breaks cmake code which tests for it.

Pick-to: 6.10
Change-Id: Ibfe0fef41c76a04de8a2989869475224eccc2658
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Piotr Wierciński <piotr.wiercinski@qt.io>
Reviewed-by: Lorn Potter <lorn.potter@qt.io>